fixup(cycle-005 PR-A3.13): iter-2 — drop tsx in publish path + pack-vs-publish gate
Bridge iter-1 verdict: COMMENT (0 BLOCKER, 0 HIGH_CONSENSUS, 2 disputed,
3 PRAISE). Two medium-severity actionable items + 3 low-severity polish:

F-001 tsx-in-publish-path:
  Renamed scripts/prepack-clean.ts → scripts/prepack-clean.mjs and
  switched the package.json hook to `node` (no tsx). Removes a publish-
  time toolchain dependency; the script uses no real TS features beyond
  `as const`. Lifecycle hooks should have minimal toolchain coupling
  (Cargo build.rs lesson).

destructive-on-dev-pack:
  Gated destructive cleanup on `process.env.npm_command === 'publish'`.
  Local `npm pack` for tarball inspection no longer silently deletes
  the developer's ~260 MB cargo cache; instead emits a "would-remove"
  dry-run report. Power-users who want a publish-equivalent pack
  tarball can run `npm_command=publish npm pack`. Tradeoff documented
  in the script header (Google SRE "toxic side effect" pattern).

F-002 no try/catch:
  Wrapped `rmSync` in try/catch with WARN logging on failure. Transient
  permission/file-lock issues no longer abort prepack and block publish.

race-statSync-rmSync (cosmetic):
  Encoded `isDir` boolean in PATHS_TO_CLEAN tuple instead of calling
  statSync at log time. Eliminates the TOCTOU between existsSync and
  statSync; logging-only code no longer can crash the publish path.

prepublishOnly-redundancy:
  Documented npm lifecycle hook ordering in the script header
  (`prepublishOnly` (tsc) → `prepack` (this file)) so future hooks
  adding build outputs into the cleaned paths know the invariant.

Empirical verification (artifacts present in both branches):
  - `npm publish --dry-run` → prepack runs in destructive mode →
    artifacts removed → tarball 1.3 MB / 2,638 files
  - `npm pack --dry-run` → prepack runs in dry-run mode →
    artifacts preserved → tarball includes them (intentional;
    publish-equivalent tarball available via npm_command=publish)

Deferred (out of scope):
  - no-test-coverage (CI tarball size budget assertion):
    legitimate cycle-006 hygiene work; PR-A3.13 stays focused on
    the immediate publish-blocker fix.
