+++ package.json
@@ -69,7 +69,7 @@
   ],
   "scripts": {
     "build": "tsc",
-    "prepack": "tsx scripts/prepack-clean.ts",
+    "prepack": "node scripts/prepack-clean.mjs",
     "prepublishOnly": "tsc",
     "schema:generate": "tsx scripts/generate-schemas.ts",
     "schema:check": "tsx scripts/check-schemas.ts",
